refactor(apps/web): move menu into header folder and a group shop folder into app directory
