+++ apps/web/src/app/(shop)/page.tsx
@@ -1,0 +1,3 @@
+export default function Page(): JSX.Element {
+    return <h1>Test</h1>;
+}
