+++ apps/web/src/app/(shop)/products/page.tsx
@@ -1,0 +1,3 @@
+export default function ProductsPage(): JSX.Element {
+    return <h1>Products Page</h1>;
+}
