--- apps/web/src/app/page.tsx
@@ -1,3 +1,0 @@
-export default function Page(): JSX.Element {
-    return <h1 className="text-red-500">Hola mundo</h1>;
-}
