+++ apps/web/src/components/common/header/header.tsx
@@ -1,8 +1,8 @@
 import { Container } from "@/components/ui/objects/container/container";
 import { Logo } from "../logo/logo";
-import { Menu } from "@/components/ui/molecules/menu/menu";
 import { getData } from "@/constants/data";
 import { AiOutlineShopping, AiOutlineUser } from "react-icons/ai";
+import { Menu } from "./menu";
 
 export const Header = async () => {
     const { menu } = await getData();
@@ -10,7 +10,6 @@ export const Header = async () => {
         <header className="">
             <Container className="flex justify-between items-center border-b  border-tertiary-400">
                 <Logo />
-
                 <Menu menu={menu} />
                 <div className="flex gap-4">
                     <AiOutlineUser className="text-primary-950" />
